Move lineage types to separate module
