+++ lineage/src/main/java/org/apache/polaris/lineage/LineageData.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.core.lineage;
+package org.apache.polaris.lineage;
 
 import java.util.Objects;
 import java.util.OptionalLong;
