+++ lineage/src/main/java/org/apache/polaris/lineage/LineageQueryRequest.java
@@ -16,10 +16,16 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.core.lineage;
+package org.apache.polaris.lineage;
 
-/** Supported query granularities for lineage lookups. */
-public enum LineageGranularity {
-  DATASET,
-  COLUMN
+import java.util.Objects;
+
+/** Request model for normalized lineage lookups. */
+public record LineageQueryRequest(
+    String nodeId, LineageDirection direction, LineageGranularity granularity) {
+  public LineageQueryRequest {
+    Objects.requireNonNull(nodeId, "nodeId must be non-null");
+    Objects.requireNonNull(direction, "direction must be non-null");
+    Objects.requireNonNull(granularity, "granularity must be non-null");
+  }
 }
