+++ lineage/src/main/java/org/apache/polaris/lineage/LineageService.java
@@ -16,17 +16,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.core.lineage;
+package org.apache.polaris.lineage;
 
-import java.util.List;
-import java.util.Objects;
-
-/** Normalized response model for lineage queries. */
-public record LineageGraph(
-    LineageNode node, List<LineageNode> upstream, List<LineageNode> downstream) {
-  public LineageGraph {
-    Objects.requireNonNull(node, "node must be non-null");
-    upstream = List.copyOf(Objects.requireNonNull(upstream, "upstream must be non-null"));
-    downstream = List.copyOf(Objects.requireNonNull(downstream, "downstream must be non-null"));
-  }
+/** Service boundary for lineage operations used by transport-layer adapters. */
+public interface LineageService {
+  LineageGraph query(LineageQueryRequest request);
 }
